+++ R/gemtc2netmeta.R
@@ -1,9 +1,10 @@
-#' Create a netmeta object from a crossnma object
+#' Create a netmeta object from a gemtc object
 #' 
 #' @description
-#' Auxiliary function to create a netmeta object from a crossnma object.
+#' Auxiliary function to create a netmeta object from a gemtc object
+#' (experimental feature).
 #' 
-#' @param x A \code{crossnma} object created with R package \bold{crossnma}.
+#' @param x A \code{gemtc} object created with R package \bold{gemtc}.
 #' @param keep.samples A logical indicating whether to keep the generated
 #'   samples.
 #' @param level The level used to calculate confidence intervals for
@@ -36,7 +37,7 @@
 #'   characters used to create unique study labels.
 #'
 #' @return
-#' A netmeta object with additional class "netmeta.crossnma".
+#' A netmeta object with additional class "netmeta.gemtc".
 #' 
 #' @author Guido Schwarzer \email{guido.schwarzer@@uniklinik-freiburg.de}
 #' 
@@ -44,49 +45,49 @@
 #' 
 #' @examples
 #' \donttest{
-#' if (requireNamespace("crossnma", quietly = TRUE)) {
-#' library("crossnma")
-#' # Create a JAGS model
-#' mod <- crossnma.model(treat, id, relapse, n, design,
-#'   prt.data = ipddata, std.data = stddata,
-#'   reference = "A", trt.effect = "random", method.bias = "naive")
-#' # Fit JAGS model
-#' set.seed(1909)
-#' fit <- crossnma(mod)
+#' if (requireNamespace("gemtc", quietly = TRUE)) {
+#' library("gemtc")
+#' # Load the example network and generate a consistency model:
+#' model <- mtc.model(smoking, type="consistency")
+#' results <- mtc.run(model, thin = 10)
+#' detach("package:gemtc")
 #'
 #' # Print results in netmeta layout
-#' cro <- crossnma2netmeta(fit)
-#' cro
+#' mtc <- gemtc2netmeta(results)
+#' mtc
 #' # SUCRAs
-#' netrank(cro)
+#' netrank(mtc)
 #' # Rankogram
-#' rankogram(cro)
+#' rankogram(mtc)
 #' }
 #' }
 #' 
-#' @export crossnma2netmeta
+#' @export gemtc2netmeta
 
-crossnma2netmeta <- function(x,
-                             keep.samples = TRUE,
-                             level = gs("level"), level.ma = x$model$level.ma,
-                             reference.group = x$model$reference,
-                             baseline.reference = gs("baseline.reference"),
-                             small.values = x$model$small.values,
-                             all.treatments = gs("all.treatments"),
-                             seq = gs("seq"),
-                             backtransf = x$model$backtransf,
-                             nchar.trts = gs("nchar.trts"),
-                             nchar.studlab = gs("nchar.studlab")) {
+gemtc2netmeta <- function(x,
+                          keep.samples = TRUE,
+                          level = gs("level"),
+                          level.ma = gs("level.ma"),
+                          reference.group =
+                            levels(x$model$network$treatments$id)[1],
+                          baseline.reference = gs("baseline.reference"),
+                          small.values = gs("small.values"),
+                          all.treatments = gs("all.treatments"),
+                          seq = gs("seq"),
+                          backtransf = gs("backtransf"),
+                          nchar.trts = gs("nchar.trts"),
+                          nchar.studlab = gs("nchar.studlab")) {
   
   #
-  # (1) Check for crossnma object and extract data / settings
+  #
+  # (1) Check for gemtc object and extract data / settings
+  #
   #
   
-  chkclass(x, "crossnma")
+  chkclass(x, "mtc.result")
   #
-  if (!is.null(x$model$covariate))
-    stop("R function crossnma2netmeta() not suitable for ",
-         "network meta-regression.",
+  if (!is.null(x$model$regressor))
+    stop("R function gemtc2netmeta() not suitable for network meta-regression.",
          call. = FALSE)
   #
   chklogical(keep.samples)
@@ -100,12 +101,34 @@ crossnma2netmeta <- function(x,
   chknumeric(nchar.trts, min = 1, length = 1)
   chknumeric(nchar.studlab, min = 1, length = 1)
   #
-  common <- x$model$trt.effect == "common"
-  random <- x$model$trt.effect == "random"
-  sm <- x$model$sm
+  random <- x$model$linearModel == "random"
+  common <- !random
   #
-  trts <- as.character(x$trt.key$trt.ini)
+  likelihood <- x$model$likelihood
+  link <- x$model$link
+  #
+  # Determine summary measure
+  #
+  assign_rules <- data.frame(
+    lik = rep(c("normal", "binom", "poisson"), c(2, 3, 1)),
+    lin = c("identity", "smd", "logit", "log", "cloglog", "log"),
+    sm = c("MD", "SMD", "OR", "RR", "HR", "IRR")
+  )
+  #
+  # Get rid of warning 'Undefined global functions or variables'
+  #
+  lik <- lin <- NULL
+  #
+  sm <- assign_rules %>% filter(lik == likelihood, lin == link) %>%
+    select(sm) %>% pull()
+  #
+  if (sm == "HR")
+    stop("Not yet implemented.")
+  #
+  trts <- levels(x$model$network$treatments$id)
   labels <- sort(trts)
+  #
+  trts.long <- x$model$network$treatments$description
   #
   if (!is.null(seq))
     seq <- setseq(seq, labels)
@@ -114,27 +137,22 @@ crossnma2netmeta <- function(x,
     if (is.numeric(seq))
       seq <- as.character(seq)
   }
+  
+  
   #
-  dat <- x$model$all.data.ad
-  trt <- dat$trt
-  n <- dat$n
-  outcome <- dat$outcome
-  study <- dat$study
-  se <- dat$se
-  
-  
   #
   # (2) Calculate matrices with treatment effects, standard errors, etc.
+  #
   #
   
   dmat <- do.call(rbind, x$samples) %>% data.frame() %>%
     select(starts_with("d."))
+  dmat <- cbind(0, dmat)
   names(dmat) <- trts
   #
-  if (random) {
-    tau <- NULL
-    tmat <- do.call(rbind, x$samples) %>% data.frame() %>% select(tau)
-  }
+  tau <- NULL
+  tmat <- do.call(rbind, x$samples) %>%
+    data.frame() %>% select(starts_with("sd."))
   #
   TE.matrix <- seTE.matrix <-
     lower.matrix <- upper.matrix <-
@@ -168,31 +186,84 @@ crossnma2netmeta <- function(x,
   
   
   #
-  # (3) Create pairwise object from crossnma data and run network meta-analysis
+  #
+  # (3) Create pairwise object from gemtc data and run network meta-analysis
+  #
   #
   
-  if (sm %in% c("OR", "RR")) {
-    suppressWarnings(
-      dat <-
-        pairwise(treat = trt, event = outcome, n = n,
-                 studlab = study, sm = sm, warn = FALSE)
-    )
+  #
+  # Arm-level data
+  #
+  dat.gemtc <- x$model$network$data.ab
+  #
+  if (!is.null(dat.gemtc)) {
+    #
+    # Get rid of warning 'Undefined global functions or variables'
+    #
+    treatment <- study <- responders <- sampleSize <- exposure <-
+      mean <- std.err <- std.dev <- NULL
+    #
+    if (sm %in% c("OR", "RR")) {
+      suppressWarnings(
+        dat <-
+          pairwise(treat = treatment, studlab = study,
+                   event = responders, n = sampleSize,
+                   data = dat.gemtc,
+                   sm = sm, warn = FALSE))
+    }
+    else if (sm == "MD") {
+      #
+      # Only means and standard errors provided
+      #
+      if (is.null(dat.gemtc$std.dev)) {
+        suppressWarnings(
+          dat <-
+            pairwise(treat = treatment,
+                     TE = mean, seTE = std.err,
+                     studlab = study,
+                     data = dat.gemtc,
+                     sm = sm, warn = FALSE))
+      }
+      else {
+        suppressWarnings(
+          dat <-
+            pairwise(treat = treatment,
+                     n = sampleSize, mean = mean, sd = std.dev,
+                     studlab = study,
+                     data = dat.gemtc,
+                     sm = sm, warn = FALSE))
+      }
+    }
+    else if (sm == "SMD") {
+      suppressWarnings(
+        dat <-
+          pairwise(treat = treatment,
+                   n = sampleSize, mean = mean, sd = std.dev,
+                   studlab = study,
+                   data = dat.gemtc,
+                   sm = sm, warn = FALSE))
+    }
+    else if (sm == "IRR") {
+      suppressWarnings(
+        dat <-
+          pairwise(treat = treatment, studlab = study,
+                   event = responders, time = exposure,
+                   data = dat.gemtc,
+                   sm = sm, warn = FALSE))
+    }
   }
-  else if (sm %in% c("MD", "SMD")) {
-    suppressWarnings(
-      dat <-
-        pairwise(treat = trt,
-                 mean = outcome, n = n, sd = se,
-                 studlab = study, sm = sm, warn = FALSE)
-    )
-  }
+  else
+    stop("R function gemtc2netmeta() not suitable for relative effect data.",
+         call. = FALSE)
   #
-  res <- suppressWarnings(netmeta(dat, common = common, random = random,
-                                  warn = FALSE))
+  res <- suppressWarnings(
+    netmeta(dat, common = common, random = random, warn = FALSE))
   
   
   #
-  # (4) Add crossnma results to netmeta object
+  #
+  # (4) Add gemtc results to netmeta object
+  #
   #
   
   res <- setNA_nma(res)
@@ -206,13 +277,13 @@ crossnma2netmeta <- function(x,
       res$samples$tau <- tmat
   }
   #
-  res$method <- "crossnma"
+  res$method <- "gemtc"
   res$method.tau <- ""
   res$m <- NA
   #
   res$level <- gs("level")
   res$level.ma <- level.ma
-  res$reference.group <- reference.group
+  res$reference.group <- setchar(reference.group, trts)
   res$baseline.reference <- baseline.reference
   res$small.values <- small.values
   res$all.treatments <- all.treatments
@@ -261,10 +332,12 @@ crossnma2netmeta <- function(x,
   
   
   #
+  #
   # (5) Return netmeta object
   #
+  #
   
-  class(res) <- c(class(res), "netmeta.crossnma")
+  class(res) <- c(class(res), "netmeta.gemtc")
   #
   res
 }